// API_Client/client/models/notification-detail-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Online Shop API
 * REST API for the online shop backend
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { NotificationActionDto } from './notification-action-dto';

export interface NotificationDetailResponseDto {
    'id': number;
    'title': string;
    'contentHtml': string;
    'imageUrl'?: string;
    'actions'?: Array<NotificationActionDto> | null;
    'isRead': boolean;
    'createdAt': string;
}